Chapter 2: fix bug in code and add tests
