+++ 02-complexity/projects/final/pubspec.yaml
@@ -7,3 +7,4 @@ environment:
 
 dev_dependencies: 
   lints: ^1.0.1
+  test: ^1.19.2
